Combine client and server package
