+++ instrumentation/ktor/ktor-2.0/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/ktor/v2_0/HttpClientInstrumentation.java
@@ -13,8 +13,8 @@ import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 import io.ktor.client.HttpClientConfig;
 import io.ktor.client.engine.HttpClientEngineConfig;
 import io.opentelemetry.api.GlobalOpenTelemetry;
-import io.opentelemetry.instrumentation.ktor.v2_0.client.KtorClientTelemetry;
-import io.opentelemetry.instrumentation.ktor.v2_0.client.KtorClientTelemetryBuilder;
+import io.opentelemetry.instrumentation.ktor.v2_0.KtorClientTelemetry;
+import io.opentelemetry.instrumentation.ktor.v2_0.KtorClientTelemetryBuilder;
 import io.opentelemetry.instrumentation.ktor.v2_0.common.internal.KtorBuilderUtil;
 import io.opentelemetry.javaagent.bootstrap.internal.AgentCommonConfig;
 import io.opentelemetry.javaagent.extension.instrumentation.TypeInstrumentation;
