+++ instrumentation/ktor/ktor-2.0/javaagent/src/test/java/io/opentelemetry/instrumentation/ktor/v2_0/KtorHttpClientTest.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.ktor.v2_0.client
+package io.opentelemetry.instrumentation.ktor.v2_0
 
 import io.ktor.client.*
 import io.opentelemetry.instrumentation.testing.junit.http.HttpClientInstrumentationExtension
@@ -14,14 +14,9 @@ class KtorHttpClientTest : AbstractKtorHttpClientTest() {
   companion object {
     @JvmStatic
     @RegisterExtension
-    private val TESTING = HttpClientInstrumentationExtension.forLibrary()
+    private val TESTING = HttpClientInstrumentationExtension.forAgent()
   }
 
   override fun HttpClientConfig<*>.installTracing() {
-    install(KtorClientTracing) {
-      setOpenTelemetry(TESTING.openTelemetry)
-      capturedRequestHeaders(TEST_REQUEST_HEADER)
-      capturedResponseHeaders(TEST_RESPONSE_HEADER)
-    }
   }
 }
