+++ instrumentation/ktor/ktor-2.0/library/src/test/kotlin/io/opentelemetry/instrumentation/ktor/v2_0/KtorHttpClientTest.kt
@@ -3,9 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.ktor.v3_0.client
+package io.opentelemetry.instrumentation.ktor.v2_0
 
 import io.ktor.client.*
+import io.opentelemetry.instrumentation.ktor.v2_0.client.KtorClientTracing
 import io.opentelemetry.instrumentation.testing.junit.http.HttpClientInstrumentationExtension
 import org.junit.jupiter.api.extension.RegisterExtension
