+++ instrumentation/ktor/ktor-3.0/javaagent/src/test/java/io/opentelemetry/instrumentation/ktor/v3_0/KtorHttpServerTest.kt
@@ -3,11 +3,12 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.ktor.v2_0.server
+package io.opentelemetry.instrumentation.ktor.v3_0
 
 import io.ktor.server.application.*
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension
 import io.opentelemetry.instrumentation.testing.junit.http.HttpServerInstrumentationExtension
+import io.opentelemetry.instrumentation.testing.junit.http.HttpServerTestOptions
 import org.junit.jupiter.api.extension.RegisterExtension
 
 class KtorHttpServerTest : AbstractKtorHttpServerTest() {
@@ -15,7 +16,7 @@ class KtorHttpServerTest : AbstractKtorHttpServerTest() {
   companion object {
     @JvmStatic
     @RegisterExtension
-    val TESTING: InstrumentationExtension = HttpServerInstrumentationExtension.forLibrary()
+    val TESTING: InstrumentationExtension = HttpServerInstrumentationExtension.forAgent()
   }
 
   override fun getTesting(): InstrumentationExtension {
@@ -23,6 +24,10 @@ class KtorHttpServerTest : AbstractKtorHttpServerTest() {
   }
 
   override fun installOpenTelemetry(application: Application) {
-    KtorTestUtil.installOpenTelemetry(application, TESTING.openTelemetry)
+  }
+
+  override fun configure(options: HttpServerTestOptions) {
+    super.configure(options)
+    options.setTestException(false)
   }
 }
