+++ instrumentation/ktor/ktor-3.0/library/src/test/kotlin/io/opentelemetry/instrumentation/ktor/v3_0/KtorHttpClientTest.kt
@@ -3,9 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.ktor.v3_0.client
+package io.opentelemetry.instrumentation.ktor.v3_0
 
 import io.ktor.client.*
+import io.opentelemetry.instrumentation.ktor.v3_0.client.KtorClientTracing
 import io.opentelemetry.instrumentation.testing.junit.http.HttpClientInstrumentationExtension
 import org.junit.jupiter.api.extension.RegisterExtension
 
@@ -14,9 +15,14 @@ class KtorHttpClientTest : AbstractKtorHttpClientTest() {
   companion object {
     @JvmStatic
     @RegisterExtension
-    private val TESTING = HttpClientInstrumentationExtension.forAgent()
+    private val TESTING = HttpClientInstrumentationExtension.forLibrary()
   }
 
   override fun HttpClientConfig<*>.installTracing() {
+    install(KtorClientTracing) {
+      setOpenTelemetry(TESTING.openTelemetry)
+      capturedRequestHeaders(TEST_REQUEST_HEADER)
+      capturedResponseHeaders(TEST_RESPONSE_HEADER)
+    }
   }
 }
