+++ instrumentation/ktor/ktor-3.0/library/src/test/kotlin/io/opentelemetry/instrumentation/ktor/v3_0/KtorHttpServerTest.kt
@@ -3,9 +3,10 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.ktor.v3_0.server
+package io.opentelemetry.instrumentation.ktor.v3_0
 
 import io.ktor.server.application.*
+import io.opentelemetry.instrumentation.ktor.v3_0.server.KtorServerTracing
 import io.opentelemetry.instrumentation.testing.junit.InstrumentationExtension
 import io.opentelemetry.instrumentation.testing.junit.http.HttpServerInstrumentationExtension
 import org.junit.jupiter.api.extension.RegisterExtension
