+++ instrumentation/ktor/ktor-3.0/testing/src/main/kotlin/io/opentelemetry/instrumentation/ktor/v3_0/AbstractKtorHttpClientTest.kt
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.instrumentation.ktor.v2_0.client
+package io.opentelemetry.instrumentation.ktor.v3_0
 
 import io.ktor.client.*
 import io.ktor.client.engine.cio.*
